pyproject.toml
[project]
name = "ripley-scrapper"
version = "0.1.0"
description = "Ripley product scraper with AI-powered grouping"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
]
